requirements.txt
# LangChain Core
langchain>=0.3.20,<1.0.0
langchain-core>=0.3.41,<1.0.0
langsmith>=0.1.125,<0.4



# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn
sentencepiece

fastapi
uvicorn 
transformers

git+https://github.com/PrithivirajDamodaran/Gramformer.git
sentencepiece


deep-translator

spacy

